Refactor pom.xml and veiculos/pom.xml
This commit refactors the pom.xml and veiculos/pom.xml files. It updates the dependencies for spring-boot-starter-data-r2dbc and spring-boot-starter-webflux. It also adds dependencies for postgresql and r2dbc-h2. Additionally, it removes the dependency for springdoc-openapi-starter-webmvc-ui. These changes were made to improve the project's dependencies and ensure compatibility with the latest versions.